Backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain
langchain-community
langchain-chroma
langchain-huggingface
unstructured
